fix: ignore dev-deps if they are missing
Projects are not required to have dev dependencies. For example, this is
the pyproject.toml output from `poetry new`:

    [tool.poetry]
    name = "tmp"
    version = "0.1.0"
    description = ""
    authors = []
    readme = "README.md"

    [tool.poetry.dependencies]
    python = "^3.12"

    [build-system]
    requires = ["poetry-core"]
    build-backend = "poetry.core.masonry.api"

This patch correctly ignores dev-deps if they are not present, and it
does so by just treating them as empty.

Tested: Used `poetry new` to generate a new toml. Confirmed that before
  this patch, we crash as reported in the bug. After this patch, we
  do not.

Closes: https://github.com/lukehsiao/poetry-udeps/issues/2
Signed-off-by: Luke Hsiao <luke@hsiao.dev>